+++ ts/src/spl/token.ts
@@ -544,7 +544,7 @@ export type SplToken = {
   ];
   accounts: [
     {
-      name: "Mint";
+      name: "mint";
       type: {
         kind: "struct";
         fields: [
@@ -576,7 +576,7 @@ export type SplToken = {
       };
     },
     {
-      name: "Token";
+      name: "token";
       type: {
         kind: "struct";
         fields: [
@@ -1150,7 +1150,7 @@ export const IDL: SplToken = {
   ],
   accounts: [
     {
-      name: "Mint",
+      name: "mint",
       type: {
         kind: "struct",
         fields: [
@@ -1182,7 +1182,7 @@ export const IDL: SplToken = {
       },
     },
     {
-      name: "Token",
+      name: "token",
       type: {
         kind: "struct",
         fields: [